doc: move 'board.inc' generation 'boards.py'
